pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=6.8.0",
    "google-cloud-firestore>=2.20.2",
    "google-generativeai>=0.8.5",
    "openai>=1.82.1",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "pypdf2>=3.0.1",
    "pyrebase4>=4.8.0",
    "streamlit>=1.45.1",
]
